requirements.txt
anyio==4.9.0
async-timeout==5.0.1
browser-cookie3==0.20.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cookies==2.2.1
cryptography==45.0.5
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lz4==4.4.4
mutagen==1.47.0
pycparser==2.22
pycryptodomex==3.23.0
python-dotenv==1.1.1
python-telegram-bot==22.2
redis==6.2.0
requests==2.32.4
services==0.1.1
sniffio==1.3.1
typing_extensions==4.14.1
urllib3==2.5.0
utils==1.0.2
yt-dlp==2025.6.30
